파일 처리 offset 변수를 int -> qint64 변경

+++ CryptokiMan/encrypt_dlg.cpp
@@ -812,7 +812,7 @@ void EncryptDlg::runFileEncrypt()
     int nPartSize = manApplet->settingsMgr()->fileReadSize();
     qint64 nReadSize = 0;
     int nLeft = 0;
-    int nOffset = 0;
+    qint64 nOffset = 0;
     int nPercent = 0;
 
     QString strSrcFile = mSrcFileText->text();
